Switch from LocalDate to LocalDateTime for timestamps

+++ src/main/java/com/example/schedulemanagement/service/ScheduleServiceImpl.java
@@ -4,29 +4,32 @@ import com.example.schedulemanagement.dto.ScheduleResponseDto;
 import com.example.schedulemanagement.entity.Schedule;
 import com.example.schedulemanagement.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class ScheduleServiceImpl implements ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
 
 
     @Override
-    public ScheduleResponseDto saveSchedule(String password, String contents, String username, LocalDate createdAt, LocalDate updatedAt) {
+    public ScheduleResponseDto saveSchedule(String password, String contents, String username, LocalDateTime createdAt, LocalDateTime updatedAt) {
         Schedule schedule = new Schedule(password, username, contents, createdAt, updatedAt);
 
         return scheduleRepository.save(schedule);
     }
 
     @Override
-    public List<ScheduleResponseDto> findAll(String username, LocalDate updatedAt) {
+    public List<ScheduleResponseDto> findAll(String username, LocalDateTime updatedAt) {
 
         return scheduleRepository.findAll(username, updatedAt);
     }
@@ -42,11 +45,12 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public ScheduleResponseDto updateSchedule(Long id, String password, String username, String contents, LocalDate updatedAt) {
+    public ScheduleResponseDto updateSchedule(Long id, String password, String username, String contents, LocalDateTime updatedAt) {
         Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
 
         //비밀번호 같은지 확인하는 로직
         if (!findSchedule.getPassword().equals(password)) {
+            log.warn("비밀번호 불일치: 입력 = {}, 실제 = {}", password, findSchedule.getPassword());
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is not same");
         }
 
@@ -57,9 +61,17 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, String password) {
         // id가 존재하는지 확인
-        scheduleRepository.findByIdOrElseThrow(id);
+        Schedule findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+
+        // 비밀번호 같은지 확인
+        if (!findSchedule.getPassword().equals(password)) {
+            log.warn("비밀번호 불일치: 입력 = {}, 실제 = {}", password, findSchedule.getPassword());
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Password is not same");
+
+        }
+
         scheduleRepository.deleteSchedule(id);
     }
 }
